Skeleton soldier animation and new stage prefab
Here's the death animation for the skeleton, I know he turns, and talked
with Chris about it before sending it. If him turning is a problem, come
get me if yall don't want to deal with moving him to fix it.

There is also lighting in the scene now to set some color that we have
planned in the stage prefab